% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_mp.R
\name{check_mp}
\alias{check_mp}
\title{Check an .mp file for problems}
\usage{
check_mp(mp)
}
\arguments{
\item{mp}{The file path to a RAMAS Metapop .mp file.}
}
\value{
If no errors cause the function's execution to stop, it will return 
  the content of the .mp file.
}
\description{
Check for possible problems with a RAMAS Metapop .mp file.
}
\keyword{internal}

