% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csquare.r, R/my_utils.r
\name{code_triplet}
\alias{code_triplet}
\alias{code_triplet}
\title{Returns a triplet C-square code}
\usage{
code_triplet(lat, lon)

code_triplet(lat, lon)
}
\arguments{
\item{lat}{A vector of the "remainder" of the lat degrees}

\item{lon}{A vector of the "remainder" of the lon degrees}

\item{lat}{A vector of the "remainder" of the lat degrees}

\item{lon}{A vector of the "remainder" of the lon degrees}
}
\value{
A triplet character vector of numbers

A triplet character vector of numbers
}
\description{
Internal function, used in \code{csquare}.

Internal function, used in \code{csquare}.
}
