% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csquare.r, R/my_utils.r
\name{csquare_resolution}
\alias{csquare_resolution}
\alias{csquare_resolution}
\title{Calculate the C-square resolution}
\usage{
csquare_resolution(csquare)

csquare_resolution(csquare)
}
\arguments{
\item{csquare}{A string of csquare codes}

\item{csquare}{A string of csquare codes}
}
\description{
Calculates the resolution of a C-square code

Calculates the resolution of a C-square code
}
