% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/facet_multiple.r
\name{facet_multiple}
\alias{facet_multiple}
\title{Use \code{facet_wrap()} over multiple pages}
\usage{
facet_multiple(plot = NULL, facets = NULL, ncol = 2, nrow = 2,
  scales = "fixed")
}
\arguments{
\item{plot}{a ggplot object}

\item{facets}{variables to facet by}

\item{ncol}{number of columns}

\item{nrow}{number of rows}

\item{scales}{should scales be fixed ("fixed", the default), free ("free"), 
or free in one dimension ("free_x", "free_y")}
}
\description{
Allows \code{ggplot2} panels to be plotted over multiple pages.
}
\examples{
\dontrun{
p <- ggplot(diamonds, aes(x = price, y = carat, color = cut)) + 
geom_point(alpha = 0.5) + 
labs(x = 'Price', y = 'Carat', title = 'Diamonds')

facet_multiple(plot = p, facets = 'color', ncol = 2, nrow = 2)
}
}
\seealso{
\code{\link{facet_wrap}}, \code{\link{facet_layout}} or 
\code{gridExtra::marrangeGrob}
}
