% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mp2shp.R
\name{mp2shp}
\alias{mp2shp}
\title{Create shapefile containing Metapop population centroids}
\usage{
mp2shp(mp, coords, outfile, start, s_p4s, t_p4s)
}
\arguments{
\item{mp}{A RAMAS Metapop .mp file containing simulation results.}

\item{coords}{An object containing population coordinates. This object can be
created by using \code{\link{mp2xy}}}

\item{outfile}{The desired output filename (including full path, and without 
file extension).}

\item{start}{The value of the first timestep. If timesteps are not in 
increments of 1, it may be best to use \code{start=1}, in which case 'time'
in the resulting shapefile's attribute table will refer to the timestep 
number.}

\item{s_p4s}{(Optional) The coordinate reference system of the source
cordinates given in \code{coords}. These can be supplied as a \code{CRS} 
object or as a proj4 string.}

\item{t_p4s}{(Optional) The target coordinate reference system to which
coordinates will be projected, if supplied. These can be supplied as a
\code{CRS} object or as a proj4 string.}
}
\value{
\code{NULL}. Three files are created: outfile.shp, outfile.shx and
  outfile.dbf.
}
\description{
Creates a shapefile (.shp, and associated .shx and .dbf files) containing a 
point representing the centroid of each population. Attributes of each point 
include the population's name (\code{pop}), and and the mean population size 
at each time step.
}
\examples{
mp <- system.file('example.mp', package='mptools')
r <- system.file('example_001.asc', package='mptools')
coords <- mp2xy(mp, r, 9.975)
tmp <- tempfile() 
mp2shp(mp, coords, tmp, start=2000) # file will be created in tempdir()
}
\seealso{
\code{\link{mp2xy}}
}
\keyword{spatial}

