\name{mptools-package}
\alias{mptools-package}
\alias{mptools}

\docType{package}

\title{RAMAS Metapop tools}
\description{
Tools to extract RAMAS Metapop metadata and simulation results, and back-transform Metapop population coordinates.
}

\details{
\tabular{ll}{
Package: \tab mptools\cr
Type: \tab Package\cr
Version: \tab 1.20\cr
Date: \tab 2014-05-04\cr
License: \tab do with it what you will.\cr
LazyLoad: \tab yes\cr
}
}

\author{
John Baumgartner\cr\cr
Maintainer: John Baumgartner <johnbb@student.unimelb.edu.au>
}
