% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mptools-package.R
\docType{package}
\name{mptools}
\alias{mptools}
\alias{mptools-package}
\title{mptools: Manipulate RAMAS Metapop .mp files}
\description{
\href{https://www.ramas.com/ramas.htm#metapop}{RAMAS Metapop} is a popular 
tool for performing spatially-explicit population viability analysis. It 
stores population model parameter settings and simulation results in plain 
text (.mp) files. RAMAS software is GUI-driven, and offers limited
flexibility with respect to extraction of results for further
analysis/visualisation in, e.g., GIS or R. The \code{mptools} package
provides a means to access, summarise and visualise model outputs.
}

