% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results.R
\name{print.mpresults}
\alias{print.mpresults}
\title{Print a metapop results object}
\usage{
\method{print}{mpresults}(x, ...)
}
\arguments{
\item{x}{An mpresults object}

\item{...}{ignored}
}
\description{
Print summary information about a RAMAS Metapop simulation.
}
\details{
The print method for \code{mpresults} objects shows only summary
  information. To see the full structure of \code{mpresults} object \code{x},
  see \code{str(x)}. Individual elements can be accessed as for a normal 
  \code{list}.
}
\seealso{
\code{\link{results}}
}

