% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knt.R
\name{knt}
\alias{knt}
\title{Plot carrying capacity and abundance trajectories}
\usage{
knt(meta, kch, filename, samelims = FALSE, plotN = FALSE, results,
  layout = c(5, 10), ...)
}
\arguments{
\item{meta}{The R object holding population info returned by \link{meta}.}

\item{kch}{The R object holding kch data returned by \link{kch}.}

\item{filename}{A character string giving the filename (with extension
".pdf") to save the plot to.}

\item{samelims}{(logical) If \code{TRUE}, the y-axis limits will be constant
across plots.}

\item{plotN}{(logical) If \code{TRUE}, mean population abundance will be
plotted (solid lines) in addition to carrying capacity (dashed lines).}

\item{results}{(required only if \code{plotN} is \code{TRUE}) The R object
holding simulation results returned by \link{results}.}

\item{layout}{The number of columns and row of plots (default = 5 columns, 10
rows).}

\item{...}{Additional arguments to \link{xyplot}.}
}
\value{
A pdf file with a plot for each population showing carrying capacity
  over time.
}
\description{
Plot each population's carrying capacity and abundance over time.
}

