% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta.R
\name{meta}
\alias{meta}
\title{Extract Metapop population metadata}
\usage{
meta(mp)
}
\arguments{
\item{mp}{A character string containing the path to a RAMAS Metapop .mp file.
E.g. file='/path/to/metapop.mp'}
}
\value{
A \code{data.frame} containing one row per population, with columns: 
  \item{popName}{The name of the population.} \item{xMetapop, yMetapop}{The spatial
  coordinates of the population centroid, in RAMAS's coordinate system (see
  \link{mp2xy}).} \item{initN}{The initial abundance.}
  \item{ddType}{The density dependence type.} \item{Rmax}{The maximum
  growth rate.} \item{K}{The initial carrying capacity.}
  \item{Ksd}{The standard deviation in K, if applicable.}
  \item{allee}{The allee parameter, if applicable.} \item{kch}{A constant, or
  reference to a .kch file describing temporal change in K.}
  \item{ddDispSourcePopN}{The parameter indicating the effect of source
  population abundance on its dispersal rate.} \item{cat1LocalMulti}{The
  local multiplier for probablity of Catastrophe 1, if applicable.}
  \item{cat1LocalProb}{The local probability of Catastrophe 1, if
  applicable.} \item{includeInTotal}{Indicates whether abundance of this 
  population is included in totals reported in simulation results.} 
  \item{stageMatrix}{The name of the stage matrix in use by the population.}
  \item{relFec}{The mean fecundity of this population, relative to those
  given in the stage matrix.} \item{relSurv}{The mean survival rates of this
  population, relative to those given in the stage matrix.}
  \item{localThr}{The abundance threshold below which the population might be
  considered dead (depending on global options specified elsewhere).}
  \item{cat2LocalMulti}{The local multiplier for probability of Catastrophe
  2, if applicable.} \item{cat2LocalProb}{The local probablity of Catastrophe
  2, if applicable.} \item{sdMatrix}{The name of the standard deviation 
  matrix in use.} \item{ddDispTargetPopK}{The target population K below which
  dispersal out of this population is reduced.} \item{tSinceCat1}{The number
  of time steps since this population last experienced Catastrophe 1.}
  \item{tSinceCat2}{The number of time steps since this population last
  experienced Catastrophe 2.} \item{relDisp}{Dispersal rates of this
  population relative to those specified by the global dispersal
  matrix/function.} \item{relVarFec}{Variation in fecundity rates of this
  population, relative to those given in the standard deviation matrix.}
  \item{relVarSurv}{Variation in survival rates of this population, relative
  to those given in the standard deviation matrix.}
}
\description{
Extract population details from RAMAS Metapop .mp files.
}
\examples{
f <- system.file('litspe.mp', package='mptools')
res <- meta(f)
head(res)
}
\seealso{
\code{\link{results}}
}

