\name{mptools-package}
\alias{mptools-package}
\alias{mptools}

\docType{package}

\title{RAMAS Metapop tools}
\description{
Tools to extract and present RAMAS Metapop metadata and simulation results, and back-transform Metapop patch coordinates.
}

\details{
\tabular{ll}{
Package: \tab mptools\cr
Type: \tab Package\cr
Version: \tab 1.3\cr
Date: \tab 2014-11-14\cr
License: \tab GPL (>= 2)\cr
LazyLoad: \tab yes\cr
}
}

\author{
John Baumgartner\cr\cr
Maintainer: John Baumgartner <johnbaums@gmail.com>
}