% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mptools-package.R
\docType{package}
\name{mptools}
\alias{mptools}
\alias{mptools-package}
\title{mptools: Manipulate RAMAS Metapop .mp files}
\description{
Extraction and presentation of metadata and simulation results from RAMAS
Metapop .mp files.
}

